backend/pyproject.toml
[tool.poetry]
name = "meetime-backend"
version = "0.1.0"
description = "BackEnd for MeeTime Solution"
authors = ["Igor Kuzmenkov <tkdomigor@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
